Crate pacaptr

pacaptr is a pacman-like syntax wrapper for many package managers.

Compatibility Table

Currently, pacaptr supports the following operations:

  Module     q   qc   qe   qi   qk   ql   qm   qo   qp   qs   qu   r   rn   rns   rs   rss   s   sc   scc   sccc   sg   si   sii   sl   ss   su   suy   sw   sy   u 
  apk.rs     *             *         *         *         *    *    *   *     *    *          *   *     *                *     *    *    *    *     *    *    *    * 
  apt.rs     *   *    *    *                   *    *         *    *   *     *    *          *   *     *           *    *     *         *    *     *    *    *      
  brew.rs    *   *         *         *                   *    *    *                         *   *     *                *     *         *    *     *    *    *      
 choco.rs    *             *                                  *    *                    *    *                          *               *    *     *                
 conda.rs    *                                 *         *         *                         *   *                      *               *    *     *                
  dnf.rs     *   *    *    *         *    *    *    *    *    *    *                         *   *     *     *     *    *     *    *    *    *     *    *    *    * 
 emerge.rs   *             *         *         *         *    *    *              *          *   *     *                *               *    *     *         *      
  pip.rs     *             *                             *    *    *                         *   *                                           *          *           
  port.rs    *   *         *         *         *         *    *    *                    *    *   *     *                *               *    *     *         *      
 scoop.rs    *             *                             *    *    *   *                     *   *     *                *               *    *     *         *      
 tlmgr.rs    *             *    *    *                             *                         *                          *          *    *    *     *              * 
 zypper.rs   *   *         *         *    *    *    *    *    *    *                    *    *   *     *           *    *          *    *    *     *    *    *    * 

Note: Some flags are “translated” so are not shown in this table, eg. -p in -Sp.

Modules

dispatch

APIs for command line argument parsing and environment detection.

error

Basic error definitions specific to this crate.

exec

APIs for spawning subprocesses and handling their results.

pm

Mapping from pacman commands to various operations of specific package managers.

print

Output messages and prompts.