Module pacaptr::dispatch

APIs for command line argument parsing and environment detection.

An overall introduction of how this module works:

clap handles command line arguments and generate a Pacaptr instance holding all the flags and options.

Config reads the configuration file (if it exists) and then merges it with the current command line arguments using Pacaptr::merge_cfg.

The correct package manager to be used will be indicated by the user (through command line arguments or config file), or, if this is not the case, automatically detected by detect_pm_str.

Pacaptr::dispatch will call the corresponding trait method, eg. .suy(), according to the combination of flags and options obtained above.

Modules

cmd

Definitions for command line argument mapping and dispatching.

config

APIs for reading pacaptr configurations from the filesystem.

Structs

Pacaptr

The command line options to be collected.

Functions

detect_pm_str

Detects the name of the package manager to be used in auto dispatch.