Enum pacaptr::dispatch::cmd::Operations

pub enum Operations {
    Query {
        c: bool,
        e: bool,
        i: u32,
        k: bool,
        l: bool,
        m: bool,
        o: bool,
        p: bool,
        s: bool,
        u: bool,
    },
    Remove {
        n: bool,
        p: bool,
        s: u32,
    },
    Sync {
        c: u32,
        g: bool,
        i: u32,
        l: bool,
        p: bool,
        s: bool,
        u: bool,
        w: bool,
        y: bool,
    },
    Update {
        p: bool,
    },
}

Variants

Query

Query the package database.

Fields of Query

c: bool

View the ChangeLog of a package if it exists.

e: bool

Restrict or filter output to explicitly installed packages.

i: u32

Display information on a given package.

k: bool

Check that all files owned by the given package(s) are present on the system.

l: bool

List all files owned by a given package.

m: bool

Restrict or filter output to packages that were not found in the sync database(s).

o: bool

Search for packages that own the specified file(s).

p: bool

Signifies that the package supplied on the command line is a file and not an entry in the database.

s: bool

Search each locally-installed package for names or descriptions that match regexp.

u: bool

Restrict or filter output to packages that are out-of-date on the local system.

Remove

Remove package(s) from the system.

Fields of Remove

n: bool

Ignore file backup designations.

p: bool

Only print the targets instead of performing the actual operation.

s: u32

Remove package(s) with all their dependencies that are no longer required.

Sync

Synchronize packages.

Fields of Sync

c: u32

Remove packages that are no longer installed from the cache as well as currently unused sync databases to free up disk space.

g: bool

Display all the members for each package group specified.

i: u32

Display information on a given sync database package.

l: bool

List all packages in the specified repositories.

p: bool

Only print the targets instead of performing the actual operation.

s: bool

Search each package in the sync databases for names or descriptions that match regexp.

u: bool

Upgrade all packages that are out-of-date Each currently-installed package will be examined and upgraded if a newer package exists.

w: bool

Retrieve all packages from the server, but do not install/upgrade anything.

y: bool

Download a fresh copy of the master package database from the server.

Update

Upgrade or add package(s) to the system and install the required dependencies from sync repositories.

Fields of Update

p: bool

Only print the targets instead of performing the actual operation.