Struct pacaptr::dispatch::Pacaptr

pub struct Pacaptr {
    ops: Operations,
    using: Option<String>,
    dry_run: bool,
    needed: bool,
    no_confirm: bool,
    no_cache: bool,
    keywords: Vec<String>,
    extra_flags: Vec<String>,
}

The command line options to be collected.

Fields
ops: Operationsusing: Option<String>

Specify the package manager to be invoked.

dry_run: bool

Perform a dry run.

needed: bool

Prevent reinstalling previously installed packages.

no_confirm: bool

Answer yes to every question.

no_cache: bool

Remove cache after installation.

keywords: Vec<String>

Package name or (sometimes) regex.

extra_flags: Vec<String>

Extra Flags passed directly to backend.

Implementations

impl Pacaptr

fn merge_cfg(&self, dotfile: Config) -> Config

Generates current Config by merging current command line arguments and options obtained with clap with the dotfile Config, which has a lower precedence.

async fn dispatch_from(&self, cfg: Config) -> Result<(), Error>

Executes the job according to the flags received and the package manager detected.

Errors

See Error for a list of possible errors.

pub async fn dispatch(&self) -> Result<(), Error>

Runs dispatch_from with automatically detected Config.

Errors

See Error for a list of possible errors.