Struct pacaptr::exec::Cmd

pub(crate) struct Cmd {
    pub sudo: bool,
    pub cmd: Vec<String>,
    pub flags: Vec<String>,
    pub kws: Vec<String>,
}

A command to be executed, provided in command-flags-keywords form.

Fields
sudo: bool

Flag indicating If a normal admin needs to run this command with sudo.

cmd: Vec<String>

The “command” part of the command string, eg. brew install.

flags: Vec<String>

The “flags” part of the command string, eg. --dry-run.

kws: Vec<String>

The “keywords” part of the command string, eg. curl fish.

Implementations

impl Cmd

pub(crate) fn new(cmd: &[impl AsRef<str>]) -> Self

Makes a new Cmd instance with the given cmd part.

pub(crate) fn with_sudo(cmd: &[impl AsRef<str>]) -> Self

Makes a new Cmd instance with the given cmd part, setting sudo to true.

pub(crate) fn flags(self, flags: &[impl AsRef<str>]) -> Self

Overrides the value of flags.

pub(crate) fn kws(self, kws: &[impl AsRef<str>]) -> Self

Overrides the value of kws.

pub(crate) fn sudo(self, sudo: bool) -> Self

Overrides the value of sudo.

fn should_sudo(&self) -> bool

Determines if this command actually needs to run with sudo -S.

If a normal admin needs to run it with sudo, and we are not root, then this is the case.

fn build(self) -> Exec

Converts a Cmd object into an Exec.

impl Cmd

pub(crate) async fn exec(self, mode: Mode) -> Result<Vec<u8>, Error>

Executes a Cmd and returns its output.

The exact behavior depends on the Mode passed in (see the definition of Mode for more info).

Errors

This function might return one of the following errors:

Error::CmdJoinError

Error::CmdNoHandleError

Error::CmdSpawnError

Error::CmdWaitError

Error::CmdStatusCodeError

Error::CmdInterruptedError

async fn exec_check_output(
self,
mute: bool,
merge: bool
) -> Result<Vec<u8>, Error>

Inner implementation of Cmd::exec_checkerr (if merge is false) and Cmd::exec_checkall (otherwise).

Errors

This function might return one of the following errors:

Error::CmdJoinError

Error::CmdNoHandleError

Error::CmdSpawnError

Error::CmdWaitError

Error::CmdStatusCodeError

Error::CmdInterruptedError

async fn exec_checkall(self, mute: bool) -> Result<Vec<u8>, Error>

Executes a Cmd and returns its stdout and stderr.

If mute is false, then normal stdout/stderr output will be printed to stdout too.

Errors

This function might return one of the following errors:

Error::CmdJoinError

Error::CmdNoHandleError

Error::CmdSpawnError

Error::CmdWaitError

Error::CmdStatusCodeError

Error::CmdInterruptedError

async fn exec_checkerr(self, mute: bool) -> Result<Vec<u8>, Error>

Executes a Cmd and collects its stderr.

If mute is false, then its stderr output will be printed to stderr too.

Errors

This function might return one of the following errors:

Error::CmdJoinError

Error::CmdNoHandleError

Error::CmdSpawnError

Error::CmdWaitError

Error::CmdStatusCodeError

Error::CmdInterruptedError

async fn exec_prompt(self, mute: bool) -> Result<Vec<u8>, Error>

Executes a Cmd and collects its stderr.

If mute is false, then its stderr output will be printed to stderr too.

This function behaves just like exec_checkerr, but in addition, the user will be prompted if (s)he wishes to continue with the command execution.

Errors

This function might return one of the following errors:

Error::CmdJoinError

Error::CmdNoHandleError

Error::CmdSpawnError

Error::CmdWaitError

Error::CmdStatusCodeError

Error::CmdInterruptedError