Trait pacaptr::pm::Pm

pub(crate) trait Pm: Sync {

    fn name(&self) -> &str;

    fn cfg(&self) -> &Config;

    fn boxed<'a>(self) -> Box<dyn Pm + 'a>
    where
        Self: Sized + 'a,
    { ... }

    fn q<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
        &'life0 self, 
        _kws: &'life1 [&'life2 str], 
        _flags: &'life3 [&'life4 str]
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait,
        'life4: 'async_trait,
        Self: 'async_trait,
    { ... }

    fn qc<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
        &'life0 self, 
        _kws: &'life1 [&'life2 str], 
        _flags: &'life3 [&'life4 str]
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait,
        'life4: 'async_trait,
        Self: 'async_trait,
    { ... }

    fn qe<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
        &'life0 self, 
        _kws: &'life1 [&'life2 str], 
        _flags: &'life3 [&'life4 str]
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait,
        'life4: 'async_trait,
        Self: 'async_trait,
    { ... }

    fn qi<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
        &'life0 self, 
        _kws: &'life1 [&'life2 str], 
        _flags: &'life3 [&'life4 str]
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait,
        'life4: 'async_trait,
        Self: 'async_trait,
    { ... }

    fn qk<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
        &'life0 self, 
        _kws: &'life1 [&'life2 str], 
        _flags: &'life3 [&'life4 str]
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait,
        'life4: 'async_trait,
        Self: 'async_trait,
    { ... }

    fn ql<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
        &'life0 self, 
        _kws: &'life1 [&'life2 str], 
        _flags: &'life3 [&'life4 str]
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait,
        'life4: 'async_trait,
        Self: 'async_trait,
    { ... }

    fn qm<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
        &'life0 self, 
        _kws: &'life1 [&'life2 str], 
        _flags: &'life3 [&'life4 str]
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait,
        'life4: 'async_trait,
        Self: 'async_trait,
    { ... }

    fn qo<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
        &'life0 self, 
        _kws: &'life1 [&'life2 str], 
        _flags: &'life3 [&'life4 str]
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait,
        'life4: 'async_trait,
        Self: 'async_trait,
    { ... }

    fn qp<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
        &'life0 self, 
        _kws: &'life1 [&'life2 str], 
        _flags: &'life3 [&'life4 str]
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait,
        'life4: 'async_trait,
        Self: 'async_trait,
    { ... }

    fn qs<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
        &'life0 self, 
        _kws: &'life1 [&'life2 str], 
        _flags: &'life3 [&'life4 str]
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait,
        'life4: 'async_trait,
        Self: 'async_trait,
    { ... }

    fn qu<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
        &'life0 self, 
        _kws: &'life1 [&'life2 str], 
        _flags: &'life3 [&'life4 str]
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait,
        'life4: 'async_trait,
        Self: 'async_trait,
    { ... }

    fn r<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
        &'life0 self, 
        _kws: &'life1 [&'life2 str], 
        _flags: &'life3 [&'life4 str]
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait,
        'life4: 'async_trait,
        Self: 'async_trait,
    { ... }

    fn rn<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
        &'life0 self, 
        _kws: &'life1 [&'life2 str], 
        _flags: &'life3 [&'life4 str]
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait,
        'life4: 'async_trait,
        Self: 'async_trait,
    { ... }

    fn rns<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
        &'life0 self, 
        _kws: &'life1 [&'life2 str], 
        _flags: &'life3 [&'life4 str]
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait,
        'life4: 'async_trait,
        Self: 'async_trait,
    { ... }

    fn rs<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
        &'life0 self, 
        _kws: &'life1 [&'life2 str], 
        _flags: &'life3 [&'life4 str]
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait,
        'life4: 'async_trait,
        Self: 'async_trait,
    { ... }

    fn rss<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
        &'life0 self, 
        _kws: &'life1 [&'life2 str], 
        _flags: &'life3 [&'life4 str]
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait,
        'life4: 'async_trait,
        Self: 'async_trait,
    { ... }

    fn s<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
        &'life0 self, 
        _kws: &'life1 [&'life2 str], 
        _flags: &'life3 [&'life4 str]
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait,
        'life4: 'async_trait,
        Self: 'async_trait,
    { ... }

    fn sc<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
        &'life0 self, 
        _kws: &'life1 [&'life2 str], 
        _flags: &'life3 [&'life4 str]
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait,
        'life4: 'async_trait,
        Self: 'async_trait,
    { ... }

    fn scc<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
        &'life0 self, 
        _kws: &'life1 [&'life2 str], 
        _flags: &'life3 [&'life4 str]
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait,
        'life4: 'async_trait,
        Self: 'async_trait,
    { ... }

    fn sccc<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
        &'life0 self, 
        _kws: &'life1 [&'life2 str], 
        _flags: &'life3 [&'life4 str]
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait,
        'life4: 'async_trait,
        Self: 'async_trait,
    { ... }

    fn sg<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
        &'life0 self, 
        _kws: &'life1 [&'life2 str], 
        _flags: &'life3 [&'life4 str]
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait,
        'life4: 'async_trait,
        Self: 'async_trait,
    { ... }

    fn si<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
        &'life0 self, 
        _kws: &'life1 [&'life2 str], 
        _flags: &'life3 [&'life4 str]
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait,
        'life4: 'async_trait,
        Self: 'async_trait,
    { ... }

    fn sii<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
        &'life0 self, 
        _kws: &'life1 [&'life2 str], 
        _flags: &'life3 [&'life4 str]
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait,
        'life4: 'async_trait,
        Self: 'async_trait,
    { ... }

    fn sl<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
        &'life0 self, 
        _kws: &'life1 [&'life2 str], 
        _flags: &'life3 [&'life4 str]
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait,
        'life4: 'async_trait,
        Self: 'async_trait,
    { ... }

    fn ss<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
        &'life0 self, 
        _kws: &'life1 [&'life2 str], 
        _flags: &'life3 [&'life4 str]
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait,
        'life4: 'async_trait,
        Self: 'async_trait,
    { ... }

    fn su<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
        &'life0 self, 
        _kws: &'life1 [&'life2 str], 
        _flags: &'life3 [&'life4 str]
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait,
        'life4: 'async_trait,
        Self: 'async_trait,
    { ... }

    fn suy<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
        &'life0 self, 
        _kws: &'life1 [&'life2 str], 
        _flags: &'life3 [&'life4 str]
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait,
        'life4: 'async_trait,
        Self: 'async_trait,
    { ... }

    fn sw<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
        &'life0 self, 
        _kws: &'life1 [&'life2 str], 
        _flags: &'life3 [&'life4 str]
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait,
        'life4: 'async_trait,
        Self: 'async_trait,
    { ... }

    fn sy<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
        &'life0 self, 
        _kws: &'life1 [&'life2 str], 
        _flags: &'life3 [&'life4 str]
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait,
        'life4: 'async_trait,
        Self: 'async_trait,
    { ... }

    fn u<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
        &'life0 self, 
        _kws: &'life1 [&'life2 str], 
        _flags: &'life3 [&'life4 str]
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait,
        'life4: 'async_trait,
        Self: 'async_trait,
    { ... }

}

The feature set of a Package Manager defined by pacman commands.

For method explanation see:

https://wiki.archlinux.org/index.php/Pacman

https://wiki.archlinux.org/index.php/Pacman/Rosetta

Required methods

fn name(&self) -> &str

Gets the name of the package manager.

fn cfg(&self) -> &Config

Gets the config of the package manager.

Provided methods

fn boxed<'a>(self) -> Box<dyn Pm + 'a> where
Self: Sized + 'a,

Wraps the Pm instance in a Box.

fn q<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
&'life0 self,
_kws: &'life1 [&'life2 str],
_flags: &'life3 [&'life4 str]
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,
'life4: 'async_trait,
Self: 'async_trait,

Q generates a list of installed packages.

fn qc<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
&'life0 self,
_kws: &'life1 [&'life2 str],
_flags: &'life3 [&'life4 str]
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,
'life4: 'async_trait,
Self: 'async_trait,

Qc shows the changelog of a package.

fn qe<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
&'life0 self,
_kws: &'life1 [&'life2 str],
_flags: &'life3 [&'life4 str]
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,
'life4: 'async_trait,
Self: 'async_trait,

Qe lists packages installed explicitly (not as dependencies).

fn qi<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
&'life0 self,
_kws: &'life1 [&'life2 str],
_flags: &'life3 [&'life4 str]
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,
'life4: 'async_trait,
Self: 'async_trait,

Qi displays local package information: name, version, description, etc.

fn qk<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
&'life0 self,
_kws: &'life1 [&'life2 str],
_flags: &'life3 [&'life4 str]
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,
'life4: 'async_trait,
Self: 'async_trait,

Qk verifies one or more packages.

fn ql<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
&'life0 self,
_kws: &'life1 [&'life2 str],
_flags: &'life3 [&'life4 str]
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,
'life4: 'async_trait,
Self: 'async_trait,

Ql displays files provided by local package.

fn qm<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
&'life0 self,
_kws: &'life1 [&'life2 str],
_flags: &'life3 [&'life4 str]
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,
'life4: 'async_trait,
Self: 'async_trait,

Qm lists packages that are installed but are not available in any installation source (anymore).

fn qo<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
&'life0 self,
_kws: &'life1 [&'life2 str],
_flags: &'life3 [&'life4 str]
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,
'life4: 'async_trait,
Self: 'async_trait,

Qo queries the package which provides FILE.

fn qp<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
&'life0 self,
_kws: &'life1 [&'life2 str],
_flags: &'life3 [&'life4 str]
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,
'life4: 'async_trait,
Self: 'async_trait,

Qp queries a package supplied through a file supplied on the command line rather than an entry in the package management database.

fn qs<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
&'life0 self,
_kws: &'life1 [&'life2 str],
_flags: &'life3 [&'life4 str]
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,
'life4: 'async_trait,
Self: 'async_trait,

Qs searches locally installed package for names or descriptions.

fn qu<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
&'life0 self,
_kws: &'life1 [&'life2 str],
_flags: &'life3 [&'life4 str]
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,
'life4: 'async_trait,
Self: 'async_trait,

Qu lists packages which have an update available.

fn r<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
&'life0 self,
_kws: &'life1 [&'life2 str],
_flags: &'life3 [&'life4 str]
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,
'life4: 'async_trait,
Self: 'async_trait,

R removes a single package, leaving all of its dependencies installed.

fn rn<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
&'life0 self,
_kws: &'life1 [&'life2 str],
_flags: &'life3 [&'life4 str]
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,
'life4: 'async_trait,
Self: 'async_trait,

Rn removes a package and skips the generation of configuration backup files.

fn rns<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
&'life0 self,
_kws: &'life1 [&'life2 str],
_flags: &'life3 [&'life4 str]
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,
'life4: 'async_trait,
Self: 'async_trait,

Rns removes a package and its dependencies which are not required by any other installed package, and skips the generation of configuration backup files.

fn rs<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
&'life0 self,
_kws: &'life1 [&'life2 str],
_flags: &'life3 [&'life4 str]
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,
'life4: 'async_trait,
Self: 'async_trait,

Rs removes a package and its dependencies which are not required by any other installed package, and not explicitly installed by the user.

fn rss<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
&'life0 self,
_kws: &'life1 [&'life2 str],
_flags: &'life3 [&'life4 str]
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,
'life4: 'async_trait,
Self: 'async_trait,

Rss removes a package and its dependencies which are not required by any other installed package.

fn s<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
&'life0 self,
_kws: &'life1 [&'life2 str],
_flags: &'life3 [&'life4 str]
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,
'life4: 'async_trait,
Self: 'async_trait,

S installs one or more packages by name.

fn sc<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
&'life0 self,
_kws: &'life1 [&'life2 str],
_flags: &'life3 [&'life4 str]
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,
'life4: 'async_trait,
Self: 'async_trait,

Sc removes all the cached packages that are not currently installed, and the unused sync database.

fn scc<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
&'life0 self,
_kws: &'life1 [&'life2 str],
_flags: &'life3 [&'life4 str]
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,
'life4: 'async_trait,
Self: 'async_trait,

Scc removes all files from the cache.

fn sccc<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
&'life0 self,
_kws: &'life1 [&'life2 str],
_flags: &'life3 [&'life4 str]
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,
'life4: 'async_trait,
Self: 'async_trait,

Sccc … What is this?

fn sg<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
&'life0 self,
_kws: &'life1 [&'life2 str],
_flags: &'life3 [&'life4 str]
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,
'life4: 'async_trait,
Self: 'async_trait,

Sg lists all packages belonging to the GROUP.

fn si<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
&'life0 self,
_kws: &'life1 [&'life2 str],
_flags: &'life3 [&'life4 str]
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,
'life4: 'async_trait,
Self: 'async_trait,

Si displays remote package information: name, version, description, etc.

fn sii<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
&'life0 self,
_kws: &'life1 [&'life2 str],
_flags: &'life3 [&'life4 str]
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,
'life4: 'async_trait,
Self: 'async_trait,

Sii displays packages which require X to be installed, aka reverse dependencies.

fn sl<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
&'life0 self,
_kws: &'life1 [&'life2 str],
_flags: &'life3 [&'life4 str]
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,
'life4: 'async_trait,
Self: 'async_trait,

Sl displays a list of all packages in all installation sources that are handled by the packages management.

fn ss<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
&'life0 self,
_kws: &'life1 [&'life2 str],
_flags: &'life3 [&'life4 str]
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,
'life4: 'async_trait,
Self: 'async_trait,

Ss searches for package(s) by searching the expression in name, description, short description.

fn su<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
&'life0 self,
_kws: &'life1 [&'life2 str],
_flags: &'life3 [&'life4 str]
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,
'life4: 'async_trait,
Self: 'async_trait,

Su updates outdated packages.

fn suy<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
&'life0 self,
_kws: &'life1 [&'life2 str],
_flags: &'life3 [&'life4 str]
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,
'life4: 'async_trait,
Self: 'async_trait,

Suy refreshes the local package database, then updates outdated packages.

fn sw<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
&'life0 self,
_kws: &'life1 [&'life2 str],
_flags: &'life3 [&'life4 str]
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,
'life4: 'async_trait,
Self: 'async_trait,

Sw retrieves all packages from the server, but does not install/upgrade anything.

fn sy<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
&'life0 self,
_kws: &'life1 [&'life2 str],
_flags: &'life3 [&'life4 str]
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,
'life4: 'async_trait,
Self: 'async_trait,

Sy refreshes the local package database.

fn u<'life0, 'life1, 'life2, 'life3, 'life4, 'async_trait>(
&'life0 self,
_kws: &'life1 [&'life2 str],
_flags: &'life3 [&'life4 str]
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,
'life4: 'async_trait,
Self: 'async_trait,

U upgrades or adds package(s) to the system and installs the required dependencies from sync repositories.