Trait pacaptr::pm::PmHelper

trait PmHelper: Pm {
    fn check_output<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        cmd: Cmd, 
        mode: PmMode, 
        strat: &'life1 Strategy
    ) -> Pin<Box<dyn Future<Output = Result<Vec<u8>, Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }

    fn run_with<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        cmd: Cmd, 
        mode: PmMode, 
        strat: &'life1 Strategy
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }

    fn run<'life0, 'async_trait>(
        &'life0 self, 
        cmd: Cmd
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
}

Extra implementation helper functions for Pm, focusing on the ability to run commands (Cmds) in a configured and Pm-specific context.

Provided methods

fn check_output<'life0, 'life1, 'async_trait>(
&'life0 self,
cmd: Cmd,
mode: PmMode,
strat: &'life1 Strategy
) -> Pin<Box<dyn Future<Output = Result<Vec<u8>, Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: Sync + 'async_trait,

Executes a command in the context of the Pm implementation. Returns the Output of this command.

fn run_with<'life0, 'life1, 'async_trait>(
&'life0 self,
cmd: Cmd,
mode: PmMode,
strat: &'life1 Strategy
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: Sync + 'async_trait,

Executes a command in the context of the Pm implementation, with custom PmMode and Strategy.

fn run<'life0, 'async_trait>(
&'life0 self,
cmd: Cmd
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Executes a command in the context of the Pm implementation with default settings.