Module pacaptr::exec

APIs for spawning subprocesses and handling their results.

Structs

Cmd

A command to be executed, provided in command-flags-keywords form.

Output

Output of running a Cmd.

Enums

Mode

Different ways in which a command shall be dealt with.

Functions

exec_tee

Takes contents from an input stream and copy to an output stream (optional) and a Vec<u8>, then returns the Vec<u8>.

grep

Finds all lines in the given text that matches all the patterns.

grep_print

Prints the result of grep line by line.

into_bytes

Turns an AsyncRead into a Stream.

is_exe

Checks if an executable exists by name (consult $PATH) or by path.

is_root

Returns true if user is root; false otherwise

prompt

Gives a prompt and returns one of the patterns matching the stdin. This action won’t end until an expected pattern is found.

Type Definitions

StatusCode

The status code type returned by a Cmd,